requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-generativeai==0.5.4  
numpy==1.26.2
faiss-cpu==1.7.4
python-dotenv==1.0.0